修改搜索聚合

+++ ly-common/src/main/java/com/leyou/common/vo/PageResult.java
@@ -2,6 +2,7 @@ package com.leyou.common.vo;
 
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
@@ -13,7 +14,7 @@ import java.util.List;
 public class PageResult<T> {
 
     private Long total;
-    private Long totalPage;
+    private Integer totalPage;
     private List<T> items;
 
     public PageResult(Long total, List<T> items) {
@@ -21,7 +22,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
